Add advanced feeder control using the feed detector in TeleOp
The feed servo is now controlled by holding down the ‘gamepad2.b’
button. The feeder servo will stop if the feed detector is tripped, or
the button is released. This means that the feeder will stop if a ball
is fed, to prevent a multi load. It can be started again by pushing the
button.